requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0